requirements.txt
fastapi
uvicorn[standard]
python-docx
pillow
fpdf
PyPDF2
requests
PdfMerger